Collect new measurements for new neighborhood.
When the new neighborhood overlaps by a lot with the previous
neighborhood, we end up computing the new step vector using
the previous measurements - which are mostly the coordinates that
are "behind" the new home coordinate. This results in our home
coordinate to be repelled by the previous measurements, without any
information about additional measurements that are "front" of the
new home coordinate.