Improve the coordinate update procedure.
* Allow None measurements and give them temporary weight of zero.
This avoids drastically pushing back the coordinate than setting
them to zero throughputs.

* Introduce back-off mechanism for the case when the new home
coordinate has None measurement. In that case, find the nearest
coordinate to the "best" coordinate we've seen so far and set
that as our new home coordinate.

* Introduce decay rates for magnitude parameter. Decay our magnitude
for each call to back-offs since that means we are now around the
edge/boundary of the valid coordinates, and therefore we want to
move more carefully.

* Introduce clipping for step vector. Since no single magnitude
is likely to work across all the models, we can just set it to some
reasonably large number and just clip the final step vector to make
sure that it doesn't shoot our new coordinate too far. This ensures
(1) we make consistent steps regardless of the model and (2) avoid
too large steps.