Integrate RunConfigMeasurement into quick algo (#470)
* Add type hints

* Create NeighborhoodData for local measurement data

* Add tests for NeighborhoodData

* Fix typo

* Update new coordinate based on the objective.

* Fix test cases

* Remove unused code

* Create a dummy measurement for None results

* Change default magnitude to 3

* Clean up

* Fix logging.

* Remove NeighborhoodData

We don't need to reset the measurement for each new neighborhood.

* Fix docstring

* Add newline

* Fix the step vector computation.

* Pretty print the config dictionary in debug mode.

* Small refactor

* Handle when both metrics are zeros.

* Scale down step vector to avoid getting too big.

* Collect new measurements for new neighborhood.

When the new neighborhood overlaps by a lot with the previous
neighborhood, we end up computing the new step vector using
the previous measurements - which are mostly the coordinates that
are "behind" the new home coordinate. This results in our home
coordinate to be repelled by the previous measurements, without any
information about additional measurements that are "front" of the
new home coordinate.

* Rename to avoid confusion

* Improve the coordinate update procedure.
Allow None measurements and give them temporary weight of zero.
This avoids drastically pushing back the coordinate than setting
them to zero throughputs.

Introduce back-off mechanism for the case when the new home
coordinate has None measurement. In that case, find the nearest
coordinate to the "best" coordinate we've seen so far and set
that as our new home coordinate.

Introduce decay rates for magnitude parameter. Decay our magnitude
for each call to back-offs since that means we are now around the
edge/boundary of the valid coordinates, and therefore we want to
move more carefully.

Introduce clipping for step vector. Since no single magnitude
is likely to work across all the models, we can just set it to some
reasonably large number and just clip the final step vector to make
sure that it doesn't shoot our new coordinate too far. This ensures
(1) we make consistent steps regardless of the model and (2) avoid
too large steps.

* Define the constants under constants.py

* Revert the change.

* Clean up the codes.

* Fix test cases

* Apply feedbacks

* Apply feedbacks

* Extract debug logging statements

* Set the values to magnitude scalar